IFAKE_WebApp/requirements.txt
# Dependências principais do projeto
Django==3.2.1
numpy==1.19.5
scipy==1.6.1
pandas==1.2.3
tensorflow==2.6.0
keras==2.6.0
Pillow==9.5.0
opencv-contrib-python==4.5.5.62
matplotlib==3.3.4
scikit-image==0.18.1
scikit-learn==0.24.1
pdf2image==1.16.0
tqdm==4.59.0

# Dependências adicionadas para a análise forense de documentos
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.2